+++ src/main/java/com/paneedah/mwc/handlers/DebugHandler.java
@@ -1,6 +1,8 @@
 package com.paneedah.mwc.handlers;
 
+import com.paneedah.mwc.MWC;
 import com.paneedah.mwc.renderer.EquipmentModelPools;
+import com.paneedah.weaponlib.PlayerItemInstanceRegistry;
 import net.minecraft.util.text.TextFormatting;
 import net.minecraftforge.client.event.RenderGameOverlayEvent;
 import net.minecraftforge.fml.common.eventhandler.EventPriority;
@@ -30,6 +32,8 @@ public class DebugHandler {
             debugScreenLeft.add(String.format("%s<MWC>%s Equipment model pool size: %s", TextFormatting.GOLD, TextFormatting.RESET, EquipmentModelPools.getSize()));
             debugScreenLeft.add(String.format("%s<MWC>%s Equipment equipped model pool size: %s", TextFormatting.GOLD, TextFormatting.RESET, EquipmentModelPools.getEquippedSize()));
             debugScreenLeft.add(String.format("%s<MWC>%s Equipment equipped armor model pool size: %s", TextFormatting.GOLD, TextFormatting.RESET, EquipmentModelPools.getEquippedArmorSize()));
+            debugScreenLeft.add("");
+            debugScreenLeft.add(String.format("%s<MWC>%s Itemstack instance cache size: %s", TextFormatting.GOLD, TextFormatting.RESET, MWC.modContext.getPlayerItemInstanceRegistry().getCacheSize()));
         }
     }
 }