+++ Changelog.md
@@ -51,6 +51,7 @@ like knowing which weapon is better or whether an attachment will be useful to y
 - Added tooltips to melee weapons
 - Added tooltips to attachments
 - Added new `enableStatusMessages` config
+- Added player item instance registry itemstack instance cache size to the F3 debug
 
 ### Changed